backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.1",
    "dj-rest-auth>=7.0.1",
    "django-allauth>=65.11.2",
    "django-cors-headers>=4.9.0",
    "django-q2>=1.8.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "waitress>=3.0.2",
    "yfinance>=0.2.48",
]
